Make searching async
